+++ common-libs/common/CHANGELOG.md
@@ -1,3 +1,6 @@
+# release 2.14.0 (2023-03-07)
+* add optional parameter `accountNumber` for wallet initialization to derive custom account `keys`/`did`
+from the root `seed`(could be used to create several ephemeral DIDs to support not traceability option).
 # release 2.13.0 (2023-02-09)
 * add an optional field `challenge` to the `validatePresentation` method;
 * bumped `vc-common`, `vc-data`, `tools-common`, `internal-api-clients`, `tools-openapi` libs versions.